+++ .idea/libraries/Dart_SDK.xml
@@ -1,0 +1,31 @@
+<component name="libraryTable">
+  <library name="Dart SDK">
+    <CLASSES>
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/_internal" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/async" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/cli" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/collection" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/concurrent" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/convert" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/core" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/developer" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/ffi" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/html" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/indexed_db" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/io" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/isolate" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/js" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/js_interop" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/js_interop_unsafe" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/js_util" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/math" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/mirrors" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/svg" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/typed_data" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/web_audio" />
+      <root url="file://$USER_HOME$/Downloads/flutter/bin/cache/dart-sdk/lib/web_gl" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
