+++ banorte_app/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import local_auth_darwin
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FLALocalAuthPlugin.register(with: registry.registrar(forPlugin: "FLALocalAuthPlugin"))
 }
